Review comment feedback.
Lots of little tweaks to the script as suggested by Menno.
